Extracted logic from modules
